Added just IP validation util
